refactor: promote EmbedPagesManager to utils file
